requirements.txt
torch
diffusers
transformers
gradio
Pillow
opencv-python
numpy
